backend/requirements.txt
# AIgal 后端依赖
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
aiofiles>=23.2.0
